DEV: Increase code coverage to 80% #32 refactoring optimize circle CI script
