+++ maxur-mserv-core/pom.xml
@@ -1020,8 +1020,10 @@
                     <includes>
                         <include>**/*Spec.*</include>
                     </includes>
+                    <!--suppress MavenModelInspection -->
                     <argLine>${surefireArgLine} -Xmx512m</argLine>
-                    <forkMode>pertest</forkMode>
+                    <forkCount>2</forkCount>
+                    <reuseForks>true</reuseForks>
                     <childDelegation>true</childDelegation>
                 </configuration>
                 <dependencies>
@@ -1050,6 +1052,7 @@
                     <encoding>${project.build.sourceEncoding}</encoding>
                     <perCoreThreadCount>true</perCoreThreadCount>
                     <threadCount>1</threadCount>
+                    <!--suppress MavenModelInspection -->
                     <argLine>${failsafeArgLine} -Xmx512m</argLine>
                 </configuration>
                 <dependencies>
