+++ maxur-mserv-core/src/main/kotlin/org/maxur/mserv/core/service/properties/PropertiesSourceJacksonImpl.kt
@@ -1,14 +1,33 @@
+@file:Suppress("unused")
+
 package org.maxur.mserv.core.service.properties
 
 import com.fasterxml.jackson.core.JsonFactory
 import com.fasterxml.jackson.databind.JsonMappingException
 import com.fasterxml.jackson.databind.JsonNode
 import com.fasterxml.jackson.databind.ObjectMapper
+import com.fasterxml.jackson.dataformat.yaml.YAMLFactory
+import org.jvnet.hk2.annotations.Service
 import org.maxur.mserv.core.service.jackson.ObjectMapperProvider
+import org.maxur.mserv.core.utils.Either
+import org.maxur.mserv.core.utils.either
 import java.io.File
 import java.io.InputStream
 import java.net.URI
 import java.nio.file.Paths
+
+
+@Service(name = "Json")
+class PropertiesFactoryJsonImpl : PropertiesFactory() {
+    override fun make(source: PropertiesSource): Either<Exception, Properties> =
+            either { PropertiesSourceJacksonImpl(JsonFactory(), "json", source) }
+}
+
+@Service(name = "Yaml")
+class PropertiesFactoryYamlImpl : PropertiesFactory() {
+    override fun make(source: PropertiesSource): Either<Exception, Properties> =
+            either { PropertiesSourceJacksonImpl(YAMLFactory(), "yaml", source) }
+}
 
 internal class PropertiesSourceJacksonImpl(
         factory: JsonFactory,
